requirements.txt
dnspython
pycryptodome